patch from Ron for Integer.valueOf
[git-p4: depot-paths = "//depot/code/antlr/main/": change = 4221]

+++ runtime/Java/src/org/antlr/runtime/tree/TreePatternParser.java
@@ -139,7 +139,7 @@ public class TreePatternParser {
 		}
 		
 		// create node
-		int treeNodeType = (Integer)wizard.getTokenType(tokenName);
+		int treeNodeType = wizard.getTokenType(tokenName);
 		if ( treeNodeType==Token.INVALID_TOKEN_TYPE ) {
 			return null;
 		}
